Clean edit link visibility, small refactors
