+++ src/components/generics/links/EditLink.vue
@@ -1,6 +1,5 @@
 <template>
   <router-link
-    v-if="$user.isLogged || showAlways"
     :to="{ name: documentType + '-edit', params: { id: document.document_id, lang: lang } }"
     @click.native="$emit('click')"
     rel="nofollow"
@@ -21,10 +20,6 @@ export default {
     lang: {
       type: String,
       required: true,
-    },
-    showAlways: {
-      type: Boolean,
-      default: false,
     },
   },
 };
