+++ src/js/view-mode-mixin.js
