+++ src/views/document/utils/boxes/CommentsBox.vue
@@ -82,12 +82,11 @@
 </template>
 
 <script>
-import viewModeMixin from '.././view-mode-mixin';
-
 import c2c from '@/js/apis/c2c';
 import forum from '@/js/apis/forum';
 import config from '@/js/config';
 import { requireDocumentProperty } from '@/js/properties-mixins';
+import viewModeMixin from '@/js/view-mode-mixin';
 
 const computeCooked = function (cooked) {
   cooked = cooked.replace(/<a class="mention" href="/g, '<a class="mention" href="' + forum.url);
